// contracts/typechain-types/PaymentRegistry.ts
export declare namespace PaymentRegistry {
  export type StatsStruct = {
    totalSentWei: BigNumberish;
    totalReceivedWei: BigNumberish;
    sentCount: BigNumberish;
    receivedCount: BigNumberish;
  };

  export type StatsStructOutput = [
    totalSentWei: bigint,
    totalReceivedWei: bigint,
    sentCount: bigint,
    receivedCount: bigint
  ] & {
    totalSentWei: bigint;
    totalReceivedWei: bigint;
    sentCount: bigint;
    receivedCount: bigint;
  };
}

export interface PaymentRegistryInterface extends Interface {
  getFunction(
    nameOrSignature:
      | "getStats"
      | "logPayment"
      | "statsOf"
      | "totalPayments"
      | "totalVolumeWei"
  ): FunctionFragment;

  getEvent(nameOrSignatureOrTopic: "PaymentLogged"): EventFragment;

  encodeFunctionData(
    functionFragment: "getStats",
    values: [AddressLike]
  ): string;
  encodeFunctionData(
    functionFragment: "logPayment",
    values: [AddressLike, BigNumberish, string, string, string]
  ): string;
  encodeFunctionData(
    functionFragment: "statsOf",
    values: [AddressLike]
  ): string;
  encodeFunctionData(
    functionFragment: "totalPayments",
    values?: undefined
  ): string;
  encodeFunctionData(
    functionFragment: "totalVolumeWei",
    values?: undefined
  ): string;

  decodeFunctionResult(functionFragment: "getStats", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "logPayment", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "statsOf", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "totalPayments",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "totalVolumeWei",
    data: BytesLike
  ): Result;
}

export namespace PaymentLoggedEvent {
  export type InputTuple = [
    from: AddressLike,
    to: AddressLike,
    amount: BigNumberish,
    fromHandle: string,
    toHandle: string,
    note: string,
    timestamp: BigNumberish
  ];
  export type OutputTuple = [
    from: string,
    to: string,
    amount: bigint,
    fromHandle: string,
    toHandle: string,
    note: string,
    timestamp: bigint
  ];
  export interface OutputObject {
    from: string;
    to: string;
    amount: bigint;
    fromHandle: string;
    toHandle: string;
    note: string;
    timestamp: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export interface PaymentRegistry extends BaseContract {
  connect(runner?: ContractRunner | null): PaymentRegistry;
  waitForDeployment(): Promise<this>;

  interface: PaymentRegistryInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  getStats: TypedContractMethod<
    [account: AddressLike],
    [PaymentRegistry.StatsStructOutput],
    "view"
  >;

  logPayment: TypedContractMethod<
    [
      to: AddressLike,
      amount: BigNumberish,
      fromHandle: string,
      toHandle: string,
      note: string
    ],
    [void],
    "nonpayable"
  >;

  statsOf: TypedContractMethod<
    [arg0: AddressLike],
    [
      [bigint, bigint, bigint, bigint] & {
        totalSentWei: bigint;
        totalReceivedWei: bigint;
        sentCount: bigint;
        receivedCount: bigint;
      }
    ],
    "view"
  >;

  totalPayments: TypedContractMethod<[], [bigint], "view">;

  totalVolumeWei: TypedContractMethod<[], [bigint], "view">;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "getStats"
  ): TypedContractMethod<
    [account: AddressLike],
    [PaymentRegistry.StatsStructOutput],
    "view"
  >;
  getFunction(
    nameOrSignature: "logPayment"
  ): TypedContractMethod<
    [
      to: AddressLike,
      amount: BigNumberish,
      fromHandle: string,
      toHandle: string,
      note: string
    ],
    [void],
    "nonpayable"
  >;
  getFunction(
    nameOrSignature: "statsOf"
  ): TypedContractMethod<
    [arg0: AddressLike],
    [
      [bigint, bigint, bigint, bigint] & {
        totalSentWei: bigint;
        totalReceivedWei: bigint;
        sentCount: bigint;
        receivedCount: bigint;
      }
    ],
    "view"
  >;
  getFunction(
    nameOrSignature: "totalPayments"
  ): TypedContractMethod<[], [bigint], "view">;
  getFunction(
    nameOrSignature: "totalVolumeWei"
  ): TypedContractMethod<[], [bigint], "view">;

  getEvent(
    key: "PaymentLogged"
  ): TypedContractEvent<
    PaymentLoggedEvent.InputTuple,
    PaymentLoggedEvent.OutputTuple,
    PaymentLoggedEvent.OutputObject
  >;

  filters: {
    "PaymentLogged(address,address,uint256,string,string,string,uint256)": TypedContractEvent<
      PaymentLoggedEvent.InputTuple,
      PaymentLoggedEvent.OutputTuple,
      PaymentLoggedEvent.OutputObject
    >;
    PaymentLogged: TypedContractEvent<
      PaymentLoggedEvent.InputTuple,
      PaymentLoggedEvent.OutputTuple,
      PaymentLoggedEvent.OutputObject
    >;
  };
}